# Faster multiple_of? method