@zuver 2018-09-17
Can anyone explain why simply reversing the ternary operands has an effect in this context?