@jacortinas 2018-09-18
@zuver I don't think changing the order of the operands is the change, but changing the type of equality checks being done and removing a method call.